Rename ecma335encoding -> ecma335
